+++ src/main/kotlin/Day1.kt
@@ -1,6 +1,7 @@
+import java.math.BigInteger
 
 class Day1 : AdventCalendarDay("input.txt") {
-    override fun part2Impl(): Int {
+    override fun part2Impl(): BigInteger {
         var amounts = mutableListOf<Int>()
         var tmp = 0
 
@@ -16,7 +17,7 @@ class Day1 : AdventCalendarDay("input.txt") {
         amounts.sort()
         amounts.reverse()
         amounts = amounts.subList(0, 3)//only get 1st item for part 1
-        return amounts.sum()
+        return amounts.sum().toBigInteger()
     }
 
     override fun part1Impl(): Int {
